+++ Tests/Blazor.InteractiveAuto.AspNetCoreIdentity/Blazor.InteractiveAuto.AspNetCoreIdentity/Startup.cs
@@ -46,7 +46,7 @@ namespace Blazor.InteractiveAuto.AspNetCoreIdentity
                                     }).AddIdentityCookies();
             var connectionString = Configuration.GetConnectionString("DefaultConnection") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
             services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlServer(connectionString));
+                                        options.UseSqlServer(connectionString));
             services.AddIdentityCore<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = true)
                                     .AddEntityFrameworkStores<ApplicationDbContext>()
                                     .AddSignInManager()
@@ -65,6 +65,7 @@ namespace Blazor.InteractiveAuto.AspNetCoreIdentity
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseWebAssemblyDebugging();
             }
             else
             {