+++ Tests/Blazor.InteractiveAuto.Jwt/Blazor.InteractiveAuto.Jwt/Startup.cs
@@ -54,6 +54,7 @@ namespace Blazor.InteractiveAuto.Jwt
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseWebAssemblyDebugging();
             }
             else
             {