+++ Tests/Blazor.InteractiveWebAssembly.AspNetCoreIdentity/Blazor.InteractiveWebAssembly.AspNetCoreIdentity/Startup.cs
@@ -56,7 +56,7 @@ namespace Blazor.InteractiveWebAssembly.AspNetCoreIdentity
                                     }).AddIdentityCookies();
             var connectionString = Configuration.GetConnectionString("DefaultConnection") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
             services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlServer(connectionString));
+                                        options.UseSqlServer(connectionString));
             services.AddIdentityCore<ApplicationUser>(options => options.SignIn.RequireConfirmedAccount = true)
                                     .AddEntityFrameworkStores<ApplicationDbContext>()
                                     .AddSignInManager()
@@ -75,6 +75,7 @@ namespace Blazor.InteractiveWebAssembly.AspNetCoreIdentity
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseWebAssemblyDebugging();
             }
             else
             {