+++ Tests/Blazor.InteractiveWebAssembly.Oidc/Blazor.InteractiveWebAssembly.Oidc/Startup.cs
@@ -66,6 +66,7 @@ namespace Blazor.InteractiveWebAssembly.Oidc
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseWebAssemblyDebugging();
             }
             else
             {