+++ Tests/MudBlazor.ExampleApp/MudBlazor.ExampleApp.Api/Startup.cs
@@ -56,6 +56,7 @@ namespace MudBlazor.ExampleApp.Api
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseWebAssemblyDebugging();
             }
             else
             {