+++ Modules/Intent.Modules.Blazor.Authentication/release-notes.md
@@ -1,6 +1,10 @@
-### Version 1.0.5
+### Version 1.0.6
 
 - Improvement: Fixed AntiForgery exception when re-logging in.
+- Improvement: Refresh token support for WASM / JWT authentication.
+
+### Version 1.0.5
+
 - Improvement: Updated module documentation to use centralized documentation site.
 
 ### Version 1.0.4